+++ app/Http/Controllers/KelompokPegawaiController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\KelompokPegawai;
+use App\Models\PendingAction;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class KelompokPegawaiController extends Controller
 {
@@ -44,7 +46,7 @@ class KelompokPegawaiController extends Controller
 
         // Redirect to the index page with a success message
         return redirect()->route('kelompok-pegawai.index')
-                         ->with('success', 'Unit Kerja berhasil ditambah');
+            ->with('success', 'Unit Kerja berhasil ditambah');
     }
 
     /**
@@ -74,17 +76,35 @@ class KelompokPegawaiController extends Controller
      */
     public function update(Request $request, KelompokPegawai $unitKerja)
     {
+        // Find the existing KelompokPegawai record
+        $data = KelompokPegawai::findorFail($unitKerja->id);
+
         // Validate the incoming request data
         $validated = $request->validate([
             'name' => 'required|string|max:255',
         ]);
+
+        // If the user is not an admin, save the delete request to the pending actions table
+        if (Auth::user()->role_id == 2) {
+            $validatedData = [
+                'id' => $data->id,
+                'validated_data'  => $validated,
+                'type' => 'KelompokPegawai',
+                'requested_by' => Auth::user()->id,
+                'requested_at' => now(),
+            ];
+
+            PendingAction::savePendingAction('update', $data->id, $validatedData);
+            return redirect()->route('kelompok-pegawai.index')
+                ->with('success', 'Kelompok Pegawai Update request submitted successfully.');
+        }
 
         // Update the existing KelompokPegawai record
         KelompokPegawai::where('id', $request->id)->update($validated);
 
         // Redirect to the index page with a success message
         return redirect()->route('kelompok-pegawai.index')
-                         ->with('success', 'Unit Kerja updated successfully.');
+            ->with('success', 'Unit Kerja updated successfully.');
     }
 
     /**
@@ -92,12 +112,29 @@ class KelompokPegawaiController extends Controller
      */
     public function destroy(Request $request)
     {
-        $data = KelompokPegawai::find($request->id);
+        // Find the KelompokPegawai record
+        $data = KelompokPegawai::findOrFail($request->id);
+
+        // If the user is not an admin, save the delete request to the pending actions table
+        if (Auth::user()->role_id == 2) {
+            $deleteData = [
+                'id' => $data->id,
+                'name' => $data->name,
+                'type' => 'KelompokPegawai',
+                'requested_by' => Auth::user()->id,
+                'requested_at' => now(),
+            ];
+
+            PendingAction::savePendingAction('delete', $data->id, $deleteData);
+            return redirect()->route('kelompok-pegawai.index')
+                ->with('success', 'Kelompok Pegawai delete request submitted successfully.');
+        }
+
         // Delete the KelompokPegawai record
         $data->delete();
 
         // Redirect to the index page with a success message
         return redirect()->route('kelompok-pegawai.index')
-                         ->with('success', 'Unit Kerja deleted successfully.');
+            ->with('success', 'Unit Kerja deleted successfully.');
     }
 }
